requirements.txt
requests
google-generativeai
google-cloud-texttospeech
python-dotenv
moviepy==1.0.3
numpy
feedparser
Pillow==9.5.0
PyYAML